+++ src/main/java/sample/cafekiosk/spring/api/service/order/dto/request/OrderCreateServiceRequest.java
@@ -1,0 +1,18 @@
+package sample.cafekiosk.spring.api.service.order.dto.request;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@NoArgsConstructor
+public class OrderCreateServiceRequest {
+    @NotEmpty(message = "상품 번호 리스트는 필수입니다.")
+    private List<String> productNumbers;
+
+    public OrderCreateServiceRequest(List<String> productNumbers) {
+        this.productNumbers = productNumbers;
+    }
+}
