+++ src/main/java/sample/cafekiosk/spring/api/service/product/ProductService.java
@@ -1,4 +1,4 @@
-package sample.cafekiosk.spring.domain.service;
+package sample.cafekiosk.spring.api.service.product;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
