+++ src/test/java/sample/cafekiosk/spring/domain/service/OrderServiceTest.java
@@ -12,6 +12,7 @@ import sample.cafekiosk.spring.domain.order.OrderRepository;
 import sample.cafekiosk.spring.domain.product.*;
 import sample.cafekiosk.spring.domain.product.dto.request.OrderCreateRequest;
 import sample.cafekiosk.spring.domain.product.dto.OrderResponse;
+import sample.cafekiosk.spring.api.service.order.OrderService;
 
 import java.time.LocalDateTime;
 import java.util.List;
